Refactoring and cleanup of torch-ddp (#1)
* scrip -> script

* change device to wrap after reaching max devices

* factors out boilerplate shared code

* checking whether float(loss) broke training

* Revert "checking whether float(loss) broke training"

This reverts commit 69bf2ecdd552a17d9861c3aabc20a884d5cbec11.

float(loss) was not what broke training

* add back in missing optimizer step

* unifies code across versions

* fixes order of args, better docs

* better docs

* fixes mixed tabs and spaces

* explain nproc_per_node parameter

* fixes formatting of log_all argument
