+++ examples/pytorch/pytorch-ddp/utils.py
@@ -1,0 +1,74 @@
+import argparse
+
+import torch.nn as nn
+
+
+def parse_args():
+    """
+    Parse arguments given to the script.
+
+    Returns:
+        The parsed argument object.
+    """
+    parser = argparse.ArgumentParser(
+        description="Run distributed data-parallel training and log with wandb.")
+    # Used for `distribution.launch`
+    parser.add_argument(
+        "--local_rank", type=int, default=-1, metavar="N", help="Local process rank."
+    )
+    parser.add_argument(
+        "--log_all",
+        action="store_true",
+        help="flag to log in all processes, otherwise only in rank0",
+    )
+    parser.add_argument(
+        "--epochs",
+        default=2,
+        type=int,
+        metavar="N",
+        help="number of total epochs to run",
+    )
+    parser.add_argument(
+        "--batch",
+        default=32,
+        type=int,
+        metavar="N",
+        help="number of data samples in one batch",
+    )
+    parser.add_argument(
+        "--entity",
+        type=str,
+        help="wandb entity",
+    )
+    parser.add_argument(
+        "--project",
+        type=str,
+        help="wandb project",
+    )
+    args = parser.parse_args()
+    return args
+
+
+class ConvNet(nn.Module):
+    def __init__(self, num_classes=10):
+        super(ConvNet, self).__init__()
+        self.layer1 = nn.Sequential(
+            nn.Conv2d(1, 16, kernel_size=5, stride=1, padding=2),
+            nn.BatchNorm2d(16),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=2, stride=2),
+        )
+        self.layer2 = nn.Sequential(
+            nn.Conv2d(16, 32, kernel_size=5, stride=1, padding=2),
+            nn.BatchNorm2d(32),
+            nn.ReLU(),
+            nn.MaxPool2d(kernel_size=2, stride=2),
+        )
+        self.fc = nn.Linear(7 * 7 * 32, num_classes)
+
+    def forward(self, x):
+        out = self.layer1(x)
+        out = self.layer2(out)
+        out = out.reshape(out.size(0), -1)
+        out = self.fc(out)
+        return out
